Refactor DailyMedicationManager to SubDailyManager
